requirements.txt
asttokens==3.0.0
decorator==5.2.1
executing==2.2.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.0.5
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
networkx==3.4.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyvis==0.3.2
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
PrettyPrintTree